requirements.txt
fastapi
uvicorn[standard]
gunicorn
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
